backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "Backend for the research agent"
authors = ["lz-chen <lingzhen.chen@inmeta.no>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
jupyter = "^1.0.0"
fake-useragent = "^1.5.1"
free-proxy = "^1.1.1"
python-dotenv = "^1.0.1"
feedparser = "^6.0.11"
pydantic = "^2.8.2"
arxiv = "^2.1.3"
pyalex = ">=0.16"
pydantic-settings = "^2.4.0"
llama-index = ">=0.13.0,<0.15.0"
click = "^8.1.7"
litellm = ">=1.0"
llama-index-llms-litellm = ">=0.1.0"
llama-index-embeddings-litellm = ">=0.1.0"
llama-parse = "*"
qdrant-client = "^1.11.0"
llama-index-vector-stores-qdrant = "*"
arxiv2text = "^0.1.14"
unstructured = "^0.15.5"
marker-pdf = ">=1.0.0"
llama-index-storage-docstore-redis = "*"
llama-index-storage-index-store-redis = "*"
pyvis = "^0.3.2"
llm-sandbox = {extras = ["docker"], version = "*"}
python-pptx = "^1.0.2"
pymupdf = "^1.24.9"
fastapi = "^0.112.2"
uvicorn = "^0.30.6"
llama-index-tools-tavily-research = "*"
mlflow = "^2.16.0"


[tool.poetry.group.test.dependencies]
pytest = ">=7.0"
pytest-asyncio = ">=0.23"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
